Revert "Simplify checkbox input widget"
This reverts commit bdc0f26f69abed0b95e34ed3881f20058270c717.

The reason we reverted this commit is:

When submitting form through POST method, the value of unchecked
checkbox field won't be sent to the server.

However, we need the value of the field to be updated and we still need
checkbox to be "checked" or "unchecked" to present the UI state.

So, when the checkbox is checked, the "turn on" value should be sent
to the server.

And when the checkbox is unchecked, the "turn off" value should still be
sent to the server. 0 for boolean field and 0 can be cast to false.